PYTHON/DJANGO/win_test/requirements.txt
asgiref==3.3.1
blessings==1.7
bpython==0.20.1
certifi==2020.12.5
chardet==3.0.4
curtsies==0.3.4
Django==2.2
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
greenlet==0.4.17
idna==2.10
importlib-metadata==3.3.0
Markdown==3.3.3
Pygments==2.7.3
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
amqp==1.4.9
anyjson==0.3.3
BeautifulSoup==3.2.1
beautifulsoup4==4.6.3
billiard==3.3.0.23
blessings==1.7
bpython==0.17.1
bs4==0.0.1
celery==3.1.26.post2
certifi==2018.8.24
chardet==3.0.4
curtsies==0.3.0
django-allauth==0.38.0
django-celery==3.2.2
django-filter==2.0.0
django-rest-auth==0.9.3
django-tls-middleware==0.1.1
djangorestframework==3.8.2
greenlet==0.4.15
gunicorn==19.6.0
idna==2.7
kombu==3.0.37
lxml==4.2.5
Markdown==3.0.1
oauthlib==2.1.0
psycopg2==2.7.4
Pygments==2.2.0
python-openid==2.2.5
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
selenium==3.14.1
six==1.11.0
typing==3.6.6
urllib3==1.23
wcwidth==0.1.7